Moved pump type to types package
