+++ src/main/kotlin/org/smolang/greenhouse/api/model/Plant.kt
@@ -1,9 +1,11 @@
 package org.smolang.greenhouse.api.model
 
+import org.smolang.greenhouse.api.types.PlantState
+
 class Plant (
     val plantId: String,
     val idealMoisture: Double,
     val moisture: Double,
-    val healthState: String
+    val healthState: PlantState
 ) {
 }
